Fixed some windows annoyances.
Of couse they would do something douchebaggy like this:

Sigh...

+++ src/model/plane.cc
@@ -4,12 +4,13 @@
 #include "src/model/plane.h"
 
 #include <math.h>
+#include <stdlib.h>
 
 namespace warhol {
 
-std::vector<float> Plane::Create(size_t width, size_t length) {
-  float fwidth = floor(width) / 2;
-  float flength = floor(length) / 2;
+std::vector<float> Plane::Create(float width, float length) {
+  float fwidth = (float)floor(width) / 2;
+  float flength = (float)floor(length) / 2;
 
   // UV is same as vertex position, so we can use the repeat pattern.
   auto vertices = std::vector<float>{
